fastapi-monitoring-assignment/app/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
prometheus-client==0.16.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
python-json-logger==2.0.6
requests==2.28.2
